+++ player.py
@@ -6,6 +6,7 @@ import random
 import time
 
 import utils
+from blacklist import BlacklistStorage
 from metareader.icecast import IcecastReader
 import config
 
@@ -126,7 +127,7 @@ class Player:
             print("Title changed to %s" % title)
 
             # If the title contains a blacklisted tag, reduce volume
-            if any(re.search(p, title, re.LOCALE) for p in config.blacklisted_tags if p.strip()):
+            if any(re.search(p, title, re.LOCALE) for p in BlacklistStorage.read_list() if p.strip()):
                 if not self._in_ad_block:
                     print('Advertisement tag detected.')
                     if config.block_mode in (config.BlockMode.REDUCE_VOLUME, config.BlockMode.REDUCE_AND_SWITCH):
@@ -203,7 +204,7 @@ class Player:
         # Reset volume level
         self.volume = config.max_volume
 
-        if len(config.blacklisted_tags) > 0:
+        if len(BlacklistStorage.read_list()) > 0:
             # TODO: Determine server type and use different reader for each
             self._meta_reader = IcecastReader(uri)
             self._meta_reader.user_agent = config.user_agent